+++ docs/lib/source.ts
@@ -1,5 +1,4 @@
 import type { InferPageType } from 'fumadocs-core/source'
-
 import { loader } from 'fumadocs-core/source'
 import { lucideIconsPlugin } from 'fumadocs-core/source/lucide-icons'
 import { blog as blogPosts, docs } from 'fumadocs-mdx:collections/server'
