+++ src/shared/components/pro/GoProModal/SectionHeading.tsx
@@ -1,0 +1,15 @@
+export function SectionHeading({ label }: { label: string }) {
+  return (
+    <div className='flex items-center gap-3 mb-10 mt-15'>
+      <div
+        className='h-px flex-1'
+        style={{ background: 'linear-gradient(to right, transparent, rgba(255,255,255,0.07))' }}
+      />
+      <span className='text-white/25 font-black uppercase tracking-[0.22em]'>{label}</span>
+      <div
+        className='h-px flex-1'
+        style={{ background: 'linear-gradient(to left, transparent, rgba(255,255,255,0.07))' }}
+      />
+    </div>
+  )
+}
